TD-3909 - Contrller refactor -DataService reference removed from the controller
