+++ DigitalLearningSolutions.Web/Controllers/RoleProfilesController/RoleProfilesController.cs
@@ -1,6 +1,5 @@
 ﻿namespace DigitalLearningSolutions.Web.Controllers.RoleProfilesController
 {
-    using DigitalLearningSolutions.Data.DataServices;
     using DigitalLearningSolutions.Web.Helpers;
     using DigitalLearningSolutions.Web.Services;
     using Microsoft.AspNetCore.Authorization;
@@ -14,21 +13,18 @@
         private readonly IRoleProfileService roleProfileService;
         private readonly ICommonService commonService;
         private readonly IFrameworkNotificationService frameworkNotificationService;
-        private readonly IConfigDataService configDataService;
         private readonly ILogger<RoleProfilesController> logger;
         private readonly IConfiguration config;
         public RoleProfilesController(
            IRoleProfileService roleProfileService,
            ICommonService commonService,
            IFrameworkNotificationService frameworkNotificationService,
-          IConfigDataService configDataService,
            ILogger<RoleProfilesController> logger,
            IConfiguration config)
         {
             this.roleProfileService = roleProfileService;
             this.commonService = commonService;
             this.frameworkNotificationService = frameworkNotificationService;
-            this.configDataService = configDataService;
             this.logger = logger;
             this.config = config;
         }
